+++ src/pk_core.c
@@ -25,145 +25,6 @@
   static const char* DOCSTRING(fn) = docstring; \
   static void fn(PKVM* vm)
 
-// Create a new module with the given [name] and returns as a Module* for
-// internal. Which will be wrapped by pkNewModule to return a pkHandle*.
-static Module* newModuleInternal(PKVM* vm, const char* name);
-
-// Adds a function to the module with the give properties and add the function
-// to the module's globals variables.
-static void moduleAddFunctionInternal(PKVM* vm, Module* module,
-                                      const char* name, pkNativeFn fptr,
-                                      int arity, const char* docstring);
-
-/*****************************************************************************/
-/* CORE PUBLIC API                                                           */
-/*****************************************************************************/
-
-#define CHECK_NULL(name) \
-  ASSERT(name != NULL, "Argument " #name " was NULL.");
-
-#define CHECK_TYPE(handle, type)                 \
-  do {                                           \
-    CHECK_NULL(handle);                          \
-    ASSERT(IS_OBJ_TYPE(handle->value, type),     \
-      "Given handle is not of type " #type "."); \
-  } while (false)
-
-PkHandle* pkNewModule(PKVM* vm, const char* name) {
-  CHECK_NULL(name);
-  Module* module = newModuleInternal(vm, name);
-  return vmNewHandle(vm, VAR_OBJ(module));
-}
-
-void pkRegisterModule(PKVM* vm, PkHandle* module) {
-  CHECK_TYPE(module, OBJ_MODULE);
-
-  Module* module_ = (Module*)AS_OBJ(module->value);
-  vmRegisterModule(vm, module_, module_->name);
-}
-
-PkHandle* pkNewClass(PKVM* vm, const char* name,
-                     PkHandle* base_class, PkHandle* module,
-                     pkNewInstanceFn new_fn,
-                     pkDeleteInstanceFn delete_fn) {
-  CHECK_NULL(module);
-  CHECK_NULL(name);
-  CHECK_TYPE(module, OBJ_MODULE);
-
-  Class* super = vm->builtin_classes[PK_OBJECT];
-  if (base_class != NULL) {
-    CHECK_TYPE(base_class, OBJ_CLASS);
-    super = (Class*)AS_OBJ(base_class->value);
-  }
-
-  Class* class_ = newClass(vm, name, (int)strlen(name),
-                           super, (Module*)AS_OBJ(module->value),
-                           NULL, NULL);
-  class_->new_fn = new_fn;
-  class_->delete_fn = delete_fn;
-
-  return vmNewHandle(vm, VAR_OBJ(class_));
-}
-
-void pkClassAddMethod(PKVM* vm, PkHandle* cls,
-                      const char* name,
-                      pkNativeFn fptr, int arity) {
-  CHECK_NULL(cls);
-  CHECK_NULL(fptr);
-  CHECK_TYPE(cls, OBJ_CLASS);
-
-  Class* class_ = (Class*)AS_OBJ(cls->value);
-
-  Function* fn = newFunction(vm, name, (int)strlen(name),
-                             class_->owner, true, NULL, NULL);
-
-  // No need to push the function to temp references of the VM
-  // since it's written to the constant pool of the module and the module
-  // won't be garbage collected (class handle has reference to the module).
-
-  Closure* method = newClosure(vm, fn);
-
-  // FIXME: name "_init" is literal everywhere.
-  if (strcmp(name, "_init") == 0) {
-    class_->ctor = method;
-
-  } else {
-    vmPushTempRef(vm, &method->_super); // method.
-    pkClosureBufferWrite(&class_->methods, vm, method);
-    vmPopTempRef(vm); // method.
-  }
-}
-
-void* pkGetSelf(const PKVM* vm) {
-  ASSERT(IS_OBJ_TYPE(vm->fiber->self, OBJ_INST), OOPS);
-  Instance* inst = (Instance*)AS_OBJ(vm->fiber->self);
-  ASSERT(inst->native != NULL, OOPS);
-  return inst->native;
-}
-
-void pkModuleAddGlobal(PKVM* vm, PkHandle* module,
-                       const char* name, PkHandle* value) {
-  CHECK_TYPE(module, OBJ_MODULE);
-  CHECK_NULL(value);
-
-  moduleAddGlobal(vm, (Module*)AS_OBJ(module->value),
-                  name, (uint32_t)strlen(name), value->value);
-}
-
-PkHandle* pkModuleGetGlobal(PKVM* vm, PkHandle* module, const char* name) {
-  CHECK_TYPE(module, OBJ_MODULE);
-  CHECK_NULL(name);
-
-  Module* module_ = (Module*)AS_OBJ(module->value);
-  int index = moduleGetGlobalIndex(module_, name, (uint32_t)strlen(name));
-  if (index == -1) return NULL;
-  return vmNewHandle(vm, module_->globals.data[index]);
-}
-
-void pkModuleAddFunction(PKVM* vm, PkHandle* module, const char* name,
-                         pkNativeFn fptr, int arity) {
-  CHECK_TYPE(module, OBJ_MODULE);
-  CHECK_NULL(fptr);
-
-  moduleAddFunctionInternal(vm, (Module*)AS_OBJ(module->value),
-                            name, fptr, arity,
-                            NULL /*TODO: Public API for function docstring.*/);
-}
-
-PkHandle* pkModuleGetMainFunction(PKVM* vm, PkHandle* module) {
-  CHECK_TYPE(module, OBJ_MODULE);
-
-  Module* _module = (Module*)AS_OBJ(module->value);
-
-  int main_index = moduleGetGlobalIndex(_module, IMPLICIT_MAIN_NAME,
-                                        (uint32_t)strlen(IMPLICIT_MAIN_NAME));
-  if (main_index == -1) return NULL;
-  ASSERT_INDEX(main_index, (int)_module->globals.count);
-  Var main_fn = _module->globals.data[main_index];
-  ASSERT(IS_OBJ_TYPE(main_fn, OBJ_CLOSURE), OOPS);
-  return vmNewHandle(vm, main_fn);
-}
-
 // A convenient macro to get the nth (1 based) argument of the current
 // function.
 #define ARG(n) (vm->fiber->ret[n])
@@ -183,172 +44,6 @@ PkHandle* pkModuleGetMainFunction(PKVM* vm, PkHandle* module) {
     VM_SET_ERROR(vm, err);     \
     RET(VAR_NULL);             \
   } while(false)
-
-// Check for errors in before calling the get arg public api function.
-#define CHECK_GET_ARG_API_ERRORS()                                          \
-  do {                                                                      \
-    ASSERT(vm->fiber != NULL,                                               \
-             "This function can only be called at runtime.");               \
-    if (arg != 0) {/* If Native setter, the value would be at fiber->ret */ \
-      ASSERT(arg > 0 && arg <= ARGC, "Invalid argument index.");            \
-    }                                                                       \
-    ASSERT(value != NULL, "Argument [value] was NULL.");                    \
-  } while (false)
-
-// Set error for incompatible type provided as an argument. (TODO: got type).
-#define ERR_INVALID_ARG_TYPE(m_type)                                        \
-do {                                                                        \
-  if (arg != 0) { /* If Native setter, arg index would be 0. */             \
-    char buff[STR_INT_BUFF_SIZE];                                           \
-    sprintf(buff, "%d", arg);                                               \
-    VM_SET_ERROR(vm, stringFormat(vm, "Expected a '$' at argument $.",      \
-                                      m_type, buff));                       \
-  } else {                                                                  \
-    VM_SET_ERROR(vm, stringFormat(vm, "Expected a '$'.", m_type));          \
-  }                                                                         \
-} while (false)
-
-int pkGetArgc(const PKVM* vm) {
-  ASSERT(vm->fiber != NULL, "This function can only be called at runtime.");
-  return ARGC;
-}
-
-bool pkCheckArgcRange(PKVM* vm, int argc, int min, int max) {
-  ASSERT(min <= max, "invalid argc range (min > max).");
-
-  if (argc < min) {
-    char buff[STR_INT_BUFF_SIZE]; sprintf(buff, "%d", min);
-    VM_SET_ERROR(vm, stringFormat(vm, "Expected at least %s argument(s).",
-                                       buff));
-    return false;
-
-  } else if (argc > max) {
-    char buff[STR_INT_BUFF_SIZE]; sprintf(buff, "%d", max);
-    VM_SET_ERROR(vm, stringFormat(vm, "Expected at most %s argument(s).",
-                                       buff));
-    return false;
-  }
-
-  return true;
-}
-
-PkVar pkGetArg(const PKVM* vm, int arg) {
-  ASSERT(vm->fiber != NULL, "This function can only be called at runtime.");
-  ASSERT(arg > 0 || arg <= ARGC, "Invalid argument index.");
-
-  return &(ARG(arg));
-}
-
-bool pkGetArgBool(PKVM* vm, int arg, bool* value) {
-  CHECK_GET_ARG_API_ERRORS();
-
-  Var val = ARG(arg);
-  *value = toBool(val);
-  return true;
-}
-
-bool pkGetArgNumber(PKVM* vm, int arg, double* value) {
-  CHECK_GET_ARG_API_ERRORS();
-
-  Var val = ARG(arg);
-  if (IS_NUM(val)) {
-    *value = AS_NUM(val);
-
-  } else if (IS_BOOL(val)) {
-    *value = AS_BOOL(val) ? 1 : 0;
-
-  } else {
-    ERR_INVALID_ARG_TYPE("number");
-    return false;
-  }
-
-  return true;
-}
-
-bool pkGetArgString(PKVM* vm, int arg, const char** value, uint32_t* length) {
-  CHECK_GET_ARG_API_ERRORS();
-
-  Var val = ARG(arg);
-  if (IS_OBJ_TYPE(val, OBJ_STRING)) {
-    String* str = (String*)AS_OBJ(val);
-    *value = str->data;
-    if (length) *length = str->length;
-
-  } else {
-    ERR_INVALID_ARG_TYPE("string");
-    return false;
-  }
-
-  return true;
-}
-
-bool pkGetArgValue(PKVM* vm, int arg, PkVarType type, PkVar* value) {
-  CHECK_GET_ARG_API_ERRORS();
-
-  Var val = ARG(arg);
-  if (pkGetValueType((PkVar)&val) != type) {
-    char buff[STR_INT_BUFF_SIZE]; sprintf(buff, "%d", arg);
-    VM_SET_ERROR(vm, stringFormat(vm, "Expected a $ at argument $.",
-                 getPkVarTypeName(type), buff));
-    return false;
-  }
-
-  *value = (PkVar)&val;
-  return true;
-}
-
-void pkReturnNull(PKVM* vm) {
-  RET(VAR_NULL);
-}
-
-void pkReturnBool(PKVM* vm, bool value) {
-  RET(VAR_BOOL(value));
-}
-
-void pkReturnNumber(PKVM* vm, double value) {
-  RET(VAR_NUM(value));
-}
-
-void pkReturnString(PKVM* vm, const char* value) {
-  RET(VAR_OBJ(newString(vm, value)));
-}
-
-void pkReturnStringLength(PKVM* vm, const char* value, size_t length) {
-  RET(VAR_OBJ(newStringLength(vm, value, (uint32_t)length)));
-}
-
-void pkReturnValue(PKVM* vm, PkVar value) {
-  RET(*(Var*)value);
-}
-
-void pkReturnHandle(PKVM* vm, PkHandle* handle) {
-  RET(handle->value);
-}
-
-const char* pkStringGetData(const PkVar value) {
-  const Var str = (*(const Var*)value);
-  ASSERT(IS_OBJ_TYPE(str, OBJ_STRING), "Value should be of type string.");
-  return ((String*)AS_OBJ(str))->data;
-}
-
-PkVar pkFiberGetReturnValue(const PkHandle* fiber) {
-  ASSERT(fiber != NULL, "Handle fiber was NULL.");
-  Var fb = fiber->value;
-  ASSERT(IS_OBJ_TYPE(fb, OBJ_FIBER), "Given handle is not a fiber");
-  Fiber* _fiber = (Fiber*)AS_OBJ(fb);
-  return (PkVar)_fiber->ret;
-}
-
-bool pkFiberIsDone(const PkHandle* fiber) {
-  ASSERT(fiber != NULL, "Handle fiber was NULL.");
-  Var fb = fiber->value;
-  ASSERT(IS_OBJ_TYPE(fb, OBJ_FIBER), "Given handle is not a fiber");
-  Fiber* _fiber = (Fiber*)AS_OBJ(fb);
-  return _fiber->state == FIBER_DONE;
-}
-
-#undef CHECK_NULL
-#undef CHECK_TYPE
 
 /*****************************************************************************/
 /* VALIDATORS                                                                */
@@ -833,7 +528,7 @@ static void initializeBuiltinFunctions(PKVM* vm) {
 /*****************************************************************************/
 
 // Create a module and add it to the vm's core modules, returns the module.
-static Module* newModuleInternal(PKVM* vm, const char* name) {
+Module* newModuleInternal(PKVM* vm, const char* name) {
 
   String* _name = newString(vm, name);
   vmPushTempRef(vm, &_name->_super); // _name
@@ -854,9 +549,9 @@ static Module* newModuleInternal(PKVM* vm, const char* name) {
 }
 
 // An internal function to add a function to the given [module].
-static void moduleAddFunctionInternal(PKVM* vm, Module* module,
-                                      const char* name, pkNativeFn fptr,
-                                      int arity, const char* docstring) {
+void moduleAddFunctionInternal(PKVM* vm, Module* module,
+                               const char* name, pkNativeFn fptr,
+                               int arity, const char* docstring) {
 
   Function* fn = newFunction(vm, name, (int)strlen(name),
                              module, true, docstring, NULL);
